Create rand_gen
Its a random number generator from the Numerical Recipes: The art of Scientific Computing 3rd edition in the section for pseudo random number generator. Since I wanted to write the highest quality pseudo-random number generator in python, well this is it. 
In my opinion the code is ugly as hell and well I haven't found a better way to write it. Curse the castings!